rename test_models to analytic
